examples/python/cloud-function-write-to-bigquery/requirements.txt
# Function dependencies, for example:
# package>=version
looker_sdk
google-api-python-client==1.7.9
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
google-cloud-bigquery